+++ bricks/http_client_handler/hooks/post_gen.dart
@@ -1,16 +1,3 @@
 import 'package:mason/mason.dart';
 
-Future<void> run(HookContext context) async {
-  context.logger.progress('Generating http_exception.dart');
-
-  // await generator.generate(
-  //   DirectoryGeneratorTarget(
-  //     Directory.current,
-  //   ),
-  //   fileConflictResolution: FileConflictResolution.overwrite,
-  //   logger: Logger(),
-  // );
-
-  context.logger.success('Genereated http_exception.dart');
-  ///
-}
+void run(HookContext context) {}