+++ bricks/http_exception/hooks/post_gen.dart
@@ -1,5 +1,12 @@
+import 'dart:io';
+
 import 'package:mason/mason.dart';
 
-void run(HookContext context) {
-  context.logger.success('Generated {{exception_file_name.snakeCase()}}.dart');
+Future<void> run(HookContext context) async {
+  final progress = context.logger.progress('Running dart format');
+  final outputPath =
+      '${Directory.current.path}/{{exception_file_name.snakeCase()}}.dart';
+  await Process.run('dart', ['format', outputPath]);
+  progress.complete();
+  context.logger.success('Formatted {{exception_file_name.snakeCase()}}.dart');
 }