+++ bricks/presentation_layer/__brick__/{{#getx}}{{name.snakeCase()}}{{/getx}}/{{name.snakeCase()}}.dart
@@ -1,3 +1,3 @@
 export 'binding/{{name.snakeCase()}}_binding.dart';
 export 'controller/{{name.snakeCase()}}_controller.dart';
-export 'view/{{name.snakeCase()}}_view.dart';
+export 'page/{{name.snakeCase()}}_page.dart';